refactor parquet module
